Add support for publishing Java libraries with Maven
This commit makes sure that Java projects and Java libraries can be published using the `maven` plugin. As the project is built with Gradle,
we only need to add additional mappings for the `implementation`, `api` and `testImplementation` configurations, because the other configurations
only make sense when building, not consuming.

Those are mapped to the following Maven scopes:

- `api` is mapped to `compile`
- `implementation` is mapped to `runtime`
- `testImplementation` is mapped to `test`

The `compileOnly` and `runtimeOnly` dependency buckets are only used from within a project, so should not appear in the generated pom file.